minor update CorblivarDie::getTuple, dropped not necessary cast

+++ src/CorblivarDie.hpp
@@ -140,13 +140,15 @@ class CorblivarDie {
 			return this->CBL.T[tuple];
 		};
 		inline int getTuple(Block const* block) const {
-			int index;
+			unsigned index;
 
-			for (index = 0; index < static_cast<int>(this->CBL.S.size()); index++) {
+			for (index = 0; index < this->CBL.S.size(); index++) {
+
 				if (block->id == this->CBL.S[index]->id) {
 					return index;
 				}
 			}
+
 			return -1;
 		};
